requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
faiss-cpu==1.7.4
numpy==1.24.3
openai==1.3.0
python-docx==0.8.11
pydantic==2.4.2
pickle-mixin==1.0.2
python-jose==3.3.0
faiss-cpu